Add gateway configuration section.
We already have some documentation on how to setup the ChirpStack MQTT
Forwarder on several gateways, but this is not supported on all
gateways. The goal is to create an entrypoint which might link to
ChirpStack MQTT Forwarder documentation, but could also document
different ways of configuring the gateway e.g. UDP forwarder or
out-of-the-box ChirpStack support provided by the gateway vendor.